requirements.txt
Flask==3.1.2
Flask-MySQLdb==2.0.0
mysql-connector-python==9.5.0
mysqlclient==2.2.7
PyMySQL==1.1.2
selenium==4.38.0
webdriver-manager==4.0.2
openpyxl==3.1.5
xlsxwriter==3.2.9
python-dotenv==1.2.1
requests==2.32.5
Werkzeug==3.1.3
Jinja2==3.1.6
